fix(package exports): fix package exports properly this time
Also, add tests to make sure that exports are working as expected

fix #829
