+++ src/index.js
@@ -1,1 +1,1 @@
-module.exports = require("./main");
+module.exports = require("./main").default;
